contrib/regex_bench/requirements.txt
google-re2==1.1.20240702
hyperscan==0.7.13
